+++ src/components/formcomponents/date/date-month-input.tsx
@@ -11,7 +11,6 @@ import Select from '@helsenorge/designsystem-react/components/Select';
 import { LanguageLocales } from '@helsenorge/core-utils/constants/languages';
 
 import { getId, isReadOnly, isRequired } from '../../../util';
-import { getValidationTextExtension } from '../../../util/extension';
 
 import {
   getMonthOptions,
@@ -34,6 +33,8 @@ import { findQuestionnaireItem } from '@/reducers/selectors';
 import { initialize } from '@/util/date-fns-utils';
 import { DateFormat } from '@/types/dateTypes';
 import { ReadOnly } from '../read-only/readOnly';
+import { shouldValidate } from '@/components/validation/utils';
+import { getErrorMessage } from '@/components/validation/rules';
 
 type DateMonthProps = QuestionnaireComponentItemProps & {
   locale: LanguageLocales.ENGLISH | LanguageLocales.NORWEGIAN;
@@ -89,16 +90,6 @@ export const DateYearMonthInput = ({
     setValue(`${idWithLinkIdAndItemIndex}-yearmonth-year`, year);
     setValue(`${idWithLinkIdAndItemIndex}-yearmonth-month`, month);
   }, []);
-
-  const getErrorText = (error: FieldError | undefined): string | undefined => {
-    if (error) {
-      const validationTextExtension = getValidationTextExtension(item);
-      if (validationTextExtension) {
-        return validationTextExtension;
-      }
-      return error.message;
-    }
-  };
 
   const getCombinedFieldError = (): FieldError | undefined => {
     const error = yearField.error || monthsField.error || undefined;
@@ -170,11 +161,11 @@ export const DateYearMonthInput = ({
 
   const { onChange: onChangeYear, ...restYear } = register(
     `${idWithLinkIdAndItemIndex}-yearmonth-year`,
-    pdf ? undefined : validationRulesYear
+    shouldValidate(item, pdf) ? validationRulesYear : undefined
   );
   const { onChange: onChangeMonth, ...restMonth } = register(
     `${idWithLinkIdAndItemIndex}-yearmonth-month`,
-    pdf ? undefined : validationRulesMonth
+    shouldValidate(item, pdf) ? validationRulesMonth : undefined
   );
 
   if (pdf || isReadOnly(item)) {
@@ -185,7 +176,7 @@ export const DateYearMonthInput = ({
     );
   }
   return (
-    <FormGroup error={getErrorText(getCombinedFieldError())} errorWrapperClassName={styles2.paddingBottom}>
+    <FormGroup error={getErrorMessage(item, getCombinedFieldError())} errorWrapperClassName={styles2.paddingBottom}>
       <ReferoLabel
         item={item}
         resources={resources}