+++ src/components/formcomponents/date/time.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect } from 'react';
 
 import { newTimeValueAsync } from '../../../actions/newValue';
-import { getValidationTextExtension } from '../../../util/extension';
 import { getId, isReadOnly, isRequired } from '../../../util/index';
 import { QuestionnaireComponentItemProps } from '@/components/createQuestionnaire/GenerateQuestionnaireComponents';
 import RenderRepeatButton from '../repeat/RenderRepeatButton';
@@ -29,6 +28,8 @@ import { initialize } from '@/util/date-fns-utils';
 import useOnAnswerChange from '@/hooks/useOnAnswerChange';
 import FormGroup from '@helsenorge/designsystem-react/components/FormGroup';
 import { ReadOnly } from '../read-only/readOnly';
+import { shouldValidate } from '@/components/validation/utils';
+import { getErrorMessage } from '@/components/validation/rules';
 
 export type Props = QuestionnaireComponentItemProps;
 
@@ -106,16 +107,6 @@ const Time = ({ id, index, path, linkId, pdf, idWithLinkIdAndItemIndex, children
     return time;
   };
 
-  const getErrorText = (error: FieldError | undefined): string | undefined => {
-    if (error) {
-      const validationTextExtension = getValidationTextExtension(item);
-      if (validationTextExtension) {
-        return validationTextExtension;
-      }
-      return error.message;
-    }
-  };
-
   function getCombinedFieldError(hoursField: FieldValues, minutesField: FieldValues): FieldError | undefined {
     const error = hoursField.error || minutesField.error || undefined;
     return error;
@@ -157,10 +148,13 @@ const Time = ({ id, index, path, linkId, pdf, idWithLinkIdAndItemIndex, children
     shouldUnregister: true,
   };
 
-  const { onChange: onChangeHours, ...restHours } = register(`${idWithLinkIdAndItemIndex}-hours`, pdf ? undefined : validationRulesHours);
+  const { onChange: onChangeHours, ...restHours } = register(
+    `${idWithLinkIdAndItemIndex}-hours`,
+    shouldValidate(item, pdf) ? validationRulesHours : undefined
+  );
   const { onChange: onChangeMinutes, ...restMinutes } = register(
     `${idWithLinkIdAndItemIndex}-minutes`,
-    pdf ? undefined : validationRulesMinutes
+    shouldValidate(item, pdf) ? validationRulesMinutes : undefined
   );
 
   if (pdf || isReadOnly(item)) {
@@ -180,7 +174,10 @@ const Time = ({ id, index, path, linkId, pdf, idWithLinkIdAndItemIndex, children
   }
   return (
     <div className="page_refero__component page_refero__component_time">
-      <FormGroup error={getErrorText(getCombinedFieldError(hoursField, minutesField))} errorWrapperClassName={styles.paddingBottom}>
+      <FormGroup
+        error={getErrorMessage(item, getCombinedFieldError(hoursField, minutesField))}
+        errorWrapperClassName={styles.paddingBottom}
+      >
         <ReferoLabel
           item={item}
           resources={resources}