+++ src/components/index.tsx
@@ -129,6 +129,7 @@ const Refero = (props: ReferoProps): JSX.Element | null => {
     const classes = [...defaultClasses];
 
     if (presentationButtonsType === PresentationButtonsType.None) {
+      console.log('presentationButtonsType', presentationButtonsType);
       classes.push('page_refero__hidden_buttons');
     } else if (presentationButtonsType === PresentationButtonsType.Sticky || (sticky && !presentationButtonsType)) {
       classes.push('page_refero__stickybar');
@@ -155,7 +156,6 @@ const Refero = (props: ReferoProps): JSX.Element | null => {
 
   const presentationButtonsType = getPresentationButtonsExtension(formDefinition.Content);
   const isStepView = shouldFormBeDisplayedAsStepView(formDefinition);
-
   return (
     <div className={getButtonClasses(presentationButtonsType, ['page_refero__content'], sticky)}>
       <div className="page_refero__messageboxes" />